Controller Comments
The Controller.h and Controller.cpp files had been fully explained, so the main logic operations controlling the main processes of the wasing machine can be explained. The main components are listed and explaind in the .h file. Detailed explanation of how they work can be found in the .cpp file!